Adyen:  Add additional data for airline and lodging
Description
-------------------------
GWS-67

This commit adds additional data for Adyen in order
to be able to send information regarding the airline and lodging.
To send this new data, it sends fields additional_data_airline and
additional_data_lodging as a GSF.

Unit test
-------------------------
Finished in 0.157969 seconds.

109 tests, 567 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

690.01 tests/s, 3589.31 assertions/s

Remote test
-------------------------
Finished in 202.708099 seconds.

134 tests, 447 assertions, 12 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
91.0448% passed

0.66 tests/s, 2.21 assertions/s

Rubocop
-------------------------
760 files inspected, no offenses detected